requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
feedparser==6.0.11
beautifulsoup4==4.12.3
apscheduler==3.10.4
jinja2==3.1.4
python-dateutil==2.9.0
